pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "docutura"
version = "0.1.0"
description = "Generic, offline-first, plugin-driven document intelligence application"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "DocTura Team", email = "info@doctura.dev"}
]
keywords = ["pdf", "document-intelligence", "excel", "conversion", "extraction"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "PySide6>=6.6.0",
    "pdfplumber>=0.10.0",
    "pypdf>=3.17.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    "pandas>=2.1.0",
    "pillow>=10.0.0",
    "pytesseract>=0.3.10",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-qt>=4.2.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]
ai = [
    "anthropic>=0.8.0",
    "openai>=1.6.0",
]

[project.scripts]
docutura = "docutura.app.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["docutura*"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
